[#30] Rename: icon 관련 컴포넌트 icons 폴더로 이동
