+++ src/components/Board/RecommendButton.tsx
@@ -4,7 +4,7 @@ import { useRouter } from 'next/navigation';
 import { useUserContext } from '@/context/UserContext';
 import useRecommendCount from '@/hooks/useRecommendCount';
 import { BoardCategory } from '@/types/board';
-import IconThumbsup from '../ui/IconThumbsup';
+import IconThumbsup from '../ui/icons/IconThumbsup';
 import {
   fetchRecommendPostsId,
   uploadRecommendPostId,
